@@ -11,7 +11,7 @@
     "creditcard",
     "ach"
   ],
-  "homepage": "http://hyperwallet.github.io/php-sdk",
+  "homepage": "https://github.com/hyperwallet/php-sdk",
   "license": "MIT",
   "authors": [
     {
@@ -20,9 +20,9 @@
     }
   ],
   "require": {
-    "php": ">=5.5.9",
+    "php": ">=5.6.0",
+    "ext-json": "*",
     "guzzlehttp/guzzle": "^6.2.1 || ^7.0.1",
-    "guzzlehttp/uri-template": "^0.2.0",
     "phpseclib/phpseclib": "^2.0.11",
     "gree/jose": "^2.2.1"
   },
@@ -33,9 +33,9 @@
     "psr-4": { "Hyperwallet\\Tests\\" : "tests/Hyperwallet/Tests", "ComposerScript\\" : "src/ComposerScript" }
   },
   "require-dev": {
-    "phpunit/phpunit": "^4.8",
-    "phake/phake": "^2.3",
-    "satooshi/php-coveralls": "^1.0"
+    "phpunit/phpunit": "^5.7 || ^7.0.0 || ^9.0",
+    "phake/phake": "^2.3 || ^4.2",
+    "php-coveralls/php-coveralls": "^2.5"
   },
   "scripts": {
     "post-install-cmd": "ComposerScript\\RsaOaep256AlgorithmInstaller::install"
