@@ -119,8 +119,8 @@ class HyperwalletVerificationDocumentCollection {
         return $this->documents;
     }
    
-    public function getIterator() : ArrayIterator {
-        return new ArrayIterator($this->documents);
+    public function getIterator() {
+        return new \ArrayIterator($this->documents);
     }
    
 }
